+++ ItemDisplayPlacementHelperUnity/Assets/IDRSJsonLoader/ModInfo/CHANGELOG.md
@@ -1,3 +1,7 @@
+**1.0.1**
+
+* Added support for skin-specific idrs. If `.idrsJson` has not empty `skinName` field, the rules will be applied only to that skin.
+
 **1.0.0**
 
 * Mod release.